// src/type.ts
export type EmployeeType = {
  id: string;
  employeeId: string;
  name: string;
  workingLocation: string;
  designation: string;
  branchCode: string;
  phoneNumber: string;
  addressId?: string;
};

export type TFlockReport = {
  id: string;
  housedBirds: number;
  fcr?: number;
  birdsStock: number;
  suggestTosell?: Boolean;
  averageBodyWight?: number;
  executiveId: string;
  totalMortality?: number;
  todayMortality: number;
  todayMortalityPercent?: number;
  totalMortalityPercentage?: number;
  bodyWeight: number;
  todayWeightGain: number;
  todayFeedEting: number;
  totalFeedEting: number;
  feedStock: number;
  diseases: string;
  condition: string;
  description: string;
  executiveName: string;
  flockNumber: string;
  age: number;
  imagesOne?: string;
  imageTwo?: string;
  imageThree?: string;
  visitedDate: Date;
  locationLink: string;
  farmId: string;
  flockId: string;
  birdsSalesStart: boolean;
  birdsSalesEnd: boolean;
  branchCode: string;
  createdAt: Date;
  updatedAt: Date;
};